fix: fixed the codes as per reviews
