+++ src/app/page.tsx
@@ -2,6 +2,24 @@ import CodeBox from "@/components/CodeBox";
 import PageWrapper from "@/components/PageWrapper";
 
 export default function Home() {
+	const codeString = `
+    // Example of a class in JavaScript
+    class Person {
+        constructor(name, age) {
+            this.name = name;
+            this.age = age;
+        }
+        
+        sayHello() {
+            return \`Hi, I'm \${this.name} and 
+                I'm \${this.age} years old.\`;
+        }
+    }
+
+    const person = new Person('John', 30);
+    console.log(person.sayHello());
+    // Output: Hi, I'm John and I'm 30 years old.
+    `;
 	return (
 		<PageWrapper title="Dashboard">
 			<div>
@@ -11,7 +29,7 @@ export default function Home() {
 				<p className="text-gray-700 dark:text-gray-300">
 					Welcome to your Next.js application with Tailwind CSS!
 				</p>
-				{/* <CodeBox /> */}
+				<CodeBox codeString={ codeString }/>
 			</div>
 		</PageWrapper>
 	);
