+++ src/types/Sidebar.ts
@@ -1,0 +1,15 @@
+export interface NavItem {
+    title: string;
+    items: NavSubItem[];
+}
+
+export interface NavSubItem {
+    navSubItemIndex?: string;
+    title: string;
+    href: string;
+    isPremium?: boolean;
+}
+
+export interface CodeBoxProps {
+    codeString: string;
+}
